+++ app/pages/glossary.vue
@@ -1,0 +1,11 @@
+<template>
+  <UContainer class="py-6 space-y-4">
+    <GlossaryTable />
+  </UContainer>
+</template>
+
+<script setup lang="ts">
+import GlossaryTable from '~/components/glossaries/GlossaryTable.vue'
+</script>
+
+